fix: all type double changed to float. feat: add package and add individual class for linear transformations. feat: add rotation method/class
