+++ src/Main.java
@@ -1,21 +1,27 @@
-import java.util.Objects;
+import Linear.Transforms.Rotation;
+import Linear.Transforms.Translate;
+import Matrices.Matrix;
+import Vectors.Vector;
 
 public class Main {
     public static void main(String[] args) {
-        Matrix A = new Matrix(2, 2, new double[]{ 5, 3, 2, 1});
-        Matrix M = new Matrix(2, 3, new double[]{21, 14, 9, 6, 15, 18});
-        Matrix systemA = new Matrix(3, 4, new double[]{ 1, 1, 2, 9, 2, 4, -3, 1, 3, 6, -5, 0 });
+        Vector vector2d = new Vector(2, new float[]{1, 0});
 
-        Vector vector2d = new Vector(2, new double[]{1, 2});
-        Vector result2d = LinearTransform.translate2D(vector2d, 5, 6);
+        Vector resultTranslate2d = Translate.translate2D(vector2d, 5, 6);
+        Vector resultRotation2d = Rotation.rotation2D(vector2d, 90);
 
+        Vector vector3d = new Vector(3, new float[]{1, 2, 3});
+        Vector resultTranslate3d = Translate.translate3D(vector3d, 5, 6, 7);
+        Vector resultRotation3dX = Rotation.rotation3DX(vector3d, 60);
+        Vector resultRotation3dY = Rotation.rotation3DY(vector3d, 60);
+        Vector resultRotation3dZ = Rotation.rotation3DZ(vector3d, 60);
 
-        Vector vector3d = new Vector(3, new double[]{1, 2, 3});
-        Vector result3d = LinearTransform.translate3D(vector3d, 5, 6, 7);
-
-
-        System.out.println(result2d);
-        System.out.println(result3d);
+        System.out.println("Translate 2D: " + resultTranslate2d);
+        System.out.println("Translate 3D: " + resultTranslate3d);
+        System.out.println("Rotation 2D:"+ resultRotation2d);
+        System.out.println("Rotation 3DX:"+ resultRotation3dX);
+        System.out.println("Rotation 3DY:"+ resultRotation3dY);
+        System.out.println("Rotation 3DZ:"+ resultRotation3dZ);
     }
